python-common/requirements.txt
kafka-python
kafka
cassandra-driver
pymongo

